[Onboarding] refactor ItemLink into separate component
