+++ src/Components/Onboarding/__stories__/Artists.story.tsx
@@ -1,6 +1,15 @@
 import { storiesOf } from "@storybook/react"
 import * as React from "react"
+import * as Relay from "react-relay"
+
+import { artsyNetworkLayer } from "../../../Relay/config"
+import PopularArtistQueryConfig from "../../../Relay/Queries/PopularArtist"
 
 import Artists from "../Steps/Artists"
 
-storiesOf("Onboarding").add("Artist Selector", () => <Artists onNextButtonPressed={() => null} />)
+function ConnectedPopularArtistSelector() {
+  Relay.injectNetworkLayer(artsyNetworkLayer())
+  return <Relay.RootContainer Component={Artists} route={new PopularArtistQueryConfig()} />
+}
+
+storiesOf("Onboarding").add("Artist Selector", () => <ConnectedPopularArtistSelector />)
